pyproject.toml
[project]
name = "repl-nix-demppython"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.0.3",
    "mysql-connector-python>=9.1.0",
    "sqlalchemy>=2.0.35",
]
